+++ core/src/lib.rs
@@ -1,7 +1,6 @@
 mod api_config;
 mod app_authorization;
 mod error;
-mod expirable_token;
 mod no_content_request;
 mod no_content_request_builder;
 mod personal_access_token;
@@ -12,7 +11,6 @@ mod response;
 pub use api_config::*;
 pub use app_authorization::*;
 pub use error::*;
-pub use expirable_token::*;
 pub use no_content_request::*;
 pub use no_content_request_builder::*;
 pub use personal_access_token::*;
