+++ types/src/expirable_token.rs
@@ -1,0 +1,4 @@
+pub trait ExpirableToken: Send + Sync {
+  // fn is_expired(&self) -> bool;
+  fn get_token(&self) -> Option<String>;
+}
