+++ types/src/lib.rs
@@ -15,3 +15,10 @@ mod pagination;
 
 #[cfg(feature = "pagination")]
 pub use pagination::*;
+
+mod expirable_token;
+
+pub use expirable_token::*;
+
+#[cfg(any(feature = "full", feature = "apps"))]
+mod installation_token;
